fix(parser): Fix `pBetweenParenthesis` requiring delimiters
Previously, `pBetweenParentesis` was calling `pSymbol` which expects to
find exactly the required string followed by a delimiter character. This
has been proven incorrect by a parse error of the string `"(+ 8 4)"`:
After the opening parenthesis, no delimiter is required, but whitespace
should stil be consumed if any is found.
For this reason, the previous `pSymbol` has been renamed to
`pSymbolStrict` (to hint towards the fact that it expects a delimiter),
and a new `pSymbol` has been added, which does not require a delimiter,
but consumes all encountered whitespace after matching the string.
`pBetweenParenthesis` has been fixed to use this revised `pSymbol`
(while showing no diff in the definition in `pBetweenParentesis, because
the function it calls has been modified)