fix(parser): Fix open brace/parenthesis suggesting adding linebreak
This tiny change has a rather meaningful impact.

Previously, when for example parsing an invalid function parameter list
such as `fun f(+` the parser would rightfully report the invalid '+'
token and then rightfully suggest a list of more expectable tokens, such
as a variable declaration or a closing parenthesis, which could both be
a valid continuation of this invalid snippet.
Problem is: it would also suggest adding a newline, which is a valid
continuation too, but is not meaningful at all to any user reading the
error.

This is done using the `hidden` megaparsec combinator which "hides" a
parser from the list of suggested items